Fix the warnings

+++ src/extractor/weightsBiasExtractor.cpp
@@ -15,7 +15,7 @@ std::vector<BiasesWeights> parseJSON(const std::string& jsonString)
     nlohmann::json data = nlohmann::json::parse(file);
     file.close();
 
-    for (int i = 1; i <= data.size() / 2; i++)
+    for (size_t i = 1; i <= data.size() / 2; i++)
     {
         BiasesWeights bw;
         bw.BiasName = "fc" + std::to_string(i) + ".bias";
@@ -48,19 +48,19 @@ void jsonValuePrinter(const std::vector<BiasesWeights>& importedWeightsAndBiases
     for (const auto& bw : importedWeightsAndBiases)
     {
         printf("Bias name: %s Biases size: %ld\n" , bw.BiasName.c_str(), bw.biases.size());
-        for (int i = 0; i < bw.biases.size(); i++)
+        for (size_t i = 0; i < bw.biases.size(); i++)
         {
-            printf("Bias %d: %f\n", i+1, bw.biases[i]);
+            printf("Bias %zu: %f\n", i+1, bw.biases[i]);
         }
         printf("\n");
         
         printf("Weights name: %s Weights size: %ld\n", bw.WeightsName.c_str(), bw.weights.size());
-        for (int i = 0; i < bw.weights.size(); i++)
+        for (size_t i = 0; i < bw.weights.size(); i++)
         {
-            printf("Row %d: %ld\n", i+1, bw.weights[i].size());
-            for (int j = 0; j < bw.weights[i].size(); j++)
+            printf("Row %zu: %ld\n", i+1, bw.weights[i].size());
+            for (size_t j = 0; j < bw.weights[i].size(); j++)
             {
-                printf("Weight %d: %.30f\n", j+1, bw.weights[i][j]);
+                printf("Weight %zu: %.30f\n", j+1, bw.weights[i][j]);
             }
             
         }
